+++ packages/canvas/breadcrumb/index.js
@@ -1,3 +1,4 @@
+import { HOOK_NAME } from '@opentiny/tiny-engine-entry'
 import CanvasBreadcrumb from './src/CanvasFooter.vue'
 import metaData from './meta'
 import useBreadcrumb from './src/useBreadcrumb'
@@ -5,5 +6,8 @@ import useBreadcrumb from './src/useBreadcrumb'
 export default {
   ...metaData,
   apis: useBreadcrumb(),
-  entry: CanvasBreadcrumb
+  entry: CanvasBreadcrumb,
+  composable: {
+    name: HOOK_NAME.useBreadcrumb
+  }
 }
