--- packages/controller/meta.js
@@ -1,3 +1,0 @@
-export default {
-  id: 'engine.controller'
-}
