+++ packages/controller/src/composable/useEditorInfo.js
