+++ packages/layout/index.js
@@ -1,13 +1,11 @@
 import component from './src/Main.vue'
 import metaData from './meta'
-import useLayout from './src/hooks/useLayout'
-import { HOOK_NAME } from '@opentiny/tiny-engine-entry'
+import { LayoutService } from './src/composable/'
 
 export default {
   ...metaData,
   component,
-  apis: useLayout(),
-  composable: {
-    name: HOOK_NAME.useLayout
-  }
+  metas: [LayoutService],
 }
+
+export { LayoutService }
