+++ packages/layout/src/composable/index.js
@@ -1,0 +1,11 @@
+import { HOOK_NAME } from '@opentiny/tiny-engine-entry'
+import useLayout from './useLayout'
+
+export const LayoutService = {
+  id: 'engine.service.layout',
+  type: 'MetaService',
+  apis: useLayout(),
+  composable: {
+    name: HOOK_NAME.useLayout
+  }
+}
