+++ packages/plugins/i18n/src/composable/index.js
@@ -1,0 +1,11 @@
+import { HOOK_NAME } from '@opentiny/tiny-engine-entry'
+import useTranslate from './useTranslate'
+
+export const TranslateService = {
+  id: 'engine.service.translate',
+  type: 'MetaService',
+  apis: useTranslate(),
+  composable: {
+    name: HOOK_NAME.useTranslate
+  }
+}
