+++ packages/plugins/i18n/src/composable/useTranslate.js
@@ -13,10 +13,10 @@
 import { reactive, ref } from 'vue'
 import { useHttp } from '@opentiny/tiny-engine-http'
 import { utils } from '@opentiny/tiny-engine-utils'
-import { isVsCodeEnv } from '../../js/environments'
+import { isVsCodeEnv } from '@opentiny/tiny-engine-controller/js/environments'
 import { constants } from '@opentiny/tiny-engine-utils'
-import { generateI18n } from '../../js/vscodeGenerateFile'
-import { PROP_DATA_TYPE } from '../../utils'
+import { generateI18n } from '@opentiny/tiny-engine-controller/js/vscodeGenerateFile'
+import { PROP_DATA_TYPE } from '@opentiny/tiny-engine-controller/utils'
 import { useResource, useCanvas } from '@opentiny/tiny-engine-entry'
 
 const { HOST_TYPE } = constants
