+++ packages/plugins/schema/src/Main.vue
@@ -43,7 +43,8 @@ import { nextTick, reactive, getCurrentInstance, onActivated, ref } from 'vue'
 import { Popover } from '@opentiny/vue'
 import { VueMonaco, CloseIcon } from '@opentiny/tiny-engine-common'
 import { useCanvas, useModal, useHistory, useNotify } from '@opentiny/tiny-engine-controller'
-import { obj2String, string2Obj, theme } from '@opentiny/tiny-engine-controller/adapter'
+import { theme } from '@opentiny/tiny-engine-controller/adapter'
+import { reactiveObj2String as obj2String, string2Obj } from '@opentiny/tiny-engine-controller/utils'
 import { iconSave, iconDownloadLink } from '@opentiny/vue-icon'
 
 export default {
