+++ packages/settings/events/src/components/AdvanceConfig.vue
@@ -76,8 +76,7 @@ import { ref, computed, reactive, watch } from 'vue'
 import { MetaInput } from '@opentiny/tiny-engine-common'
 import { CodeConfigurator, SwitchConfigurator, VariableConfigurator } from '@opentiny/tiny-engine-configurator'
 import { useProperties, useCanvas } from '@opentiny/tiny-engine-controller'
-import { PROP_DATA_TYPE } from '@opentiny/tiny-engine-controller/utils'
-import { string2Obj } from '@opentiny/tiny-engine-controller/adapter'
+import { PROP_DATA_TYPE, string2Obj } from '@opentiny/tiny-engine-controller/utils'
 import { constants } from '@opentiny/tiny-engine-utils'
 import { Tooltip } from '@opentiny/vue'
