+++ packages/settings/props/src/components/inputs/CodeEditor.vue
@@ -37,7 +37,8 @@
 import { reactive, getCurrentInstance, h, ref, computed } from 'vue'
 import { VueMonaco } from '@opentiny/tiny-engine-common'
 import { Button, Collapse, CollapseItem, DialogBox } from '@opentiny/vue'
-import { getExample, useCanvas, useProperties } from '@opentiny/tiny-engine-controller'
+import { useCanvas, useProperties } from '@opentiny/tiny-engine-controller'
+import { getExample } from '@opentiny/tiny-engine-controller/js/example'
 import { theme } from '@opentiny/tiny-engine-controller/adapter'
 
 export default {
