+++ packages/toolbars/generate-vue/src/composable/index.js
@@ -1,0 +1,11 @@
+import { HOOK_NAME } from '@opentiny/tiny-engine-entry'
+import useSaveLocal from './useSaveLocal'
+
+export const SaveLocalService = {
+  id: 'engine.service.savelocal',
+  type: 'MetaService',
+  apis: useSaveLocal(),
+  composable: {
+    name: HOOK_NAME.useSaveLocal
+  }
+}
