+++ packages/toolbars/redoundo/index.js
@@ -12,8 +12,12 @@
 
 import entry from './src/Main.vue'
 import metaData from './meta'
+import { HistoryService } from './src/composable/'
 
 export default {
   ...metaData,
-  entry
+  entry,
+  metas: [HistoryService]
 }
+
+export { HistoryService }
