+++ packages/toolbars/save/index.js
@@ -12,7 +12,6 @@
 
 import entry, { api } from './src/Main.vue'
 import metaData from './meta'
-import './src/hooks/useSaveLocal'
 
 export default {
   ...metaData,
